Remove old entrypoint.js and use Toolkit.run
